Add option to load stdout and stderr when requesting job status

+++ lib/galaxy/webapps/galaxy/api/job_files.py
@@ -104,7 +104,11 @@ class JobFilesAPIController(BaseGalaxyAPIController):
         target_dir = os.path.dirname(path)
         util.safe_makedirs(target_dir)
         try:
-            shutil.move(input_file.name, path)
+            if os.path.exists(path):
+                with open(path, "ab") as destination:
+                    shutil.copyfileobj(open(input_file.name, "rb"), destination)
+            else:
+                shutil.move(input_file.name, path)
         finally:
             try:
                 input_file.close()
